social_network/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
backports-zoneinfo==0.2.1
    # via django
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   pip-compile-multi
    #   pip-tools
coreapi==2.3.3
    # via
    #   -r requirements.in
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r requirements.in
    #   coreapi
    #   drf-yasg
django==4.0.4
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.2.0
    # via -r requirements.in
drf-yasg==1.20.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.1
    # via jinja2
packaging==21.3
    # via drf-yasg
pep517==0.12.0
    # via pip-tools
pip-compile-multi==2.4.5
    # via -r requirements.in
pip-tools==6.6.2
    # via pip-compile-multi
psycopg2==2.9.3
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
pyjwt==2.4.0
    # via djangorestframework-simplejwt
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via djangorestframework
requests==2.27.1
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via
    #   -r requirements.in
    #   django
tomli==2.0.1
    # via pep517
toposort==1.7
    # via pip-compile-multi
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.9
    # via requests
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
